refactor components, check credits on connect
