+++ .github/workflows/firebase-hosting-pull-request.yml
@@ -1,16 +1,16 @@
 # This file was auto-generated by the Firebase CLI
 # https://github.com/firebase/firebase-tools
 
-# name: Deploy to Firebase Hosting on PR
-# 'on': pull_request
-# jobs:
-#   build_and_preview:
-#     if: '${{ github.event.pull_request.head.repo.full_name == github.repository }}'
-#     runs-on: ubuntu-latest
-#     steps:
-#       - uses: actions/checkout@v2
-#       - uses: FirebaseExtended/action-hosting-deploy@v0
-#         with:
-#           repoToken: '${{ secrets.GITHUB_TOKEN }}'
-#           firebaseServiceAccount: '${{ secrets.FIREBASE_SERVICE_ACCOUNT_SDXCRYPTO_ALGOVERA }}'
-#           projectId: sdxcrypto-algovera
+name: Deploy to Firebase Hosting on PR
+'on': pull_request
+jobs:
+  build_and_preview:
+    if: '${{ github.event.pull_request.head.repo.full_name == github.repository }}'
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: FirebaseExtended/action-hosting-deploy@v0
+        with:
+          repoToken: '${{ secrets.GITHUB_TOKEN }}'
+          firebaseServiceAccount: '${{ secrets.FIREBASE_SERVICE_ACCOUNT_SDXCRYPTO_ALGOVERA }}'
+          projectId: sdxcrypto-algovera
