+++ frontend/components/index/hero.tsx
