+++ frontend/lib/types.tsx
@@ -1,8 +1,8 @@
 export type PageProps = {
   uid: string;
   setUid: React.Dispatch<React.SetStateAction<string>>;
-  credits: number;
-  setCredits: React.Dispatch<React.SetStateAction<number>>;
+  credits: number | null;
+  moralisAuth: any;
   creditsModalTrigger: boolean;
   setCreditsModalTrigger: React.Dispatch<React.SetStateAction<boolean>>;
 };
